Allow ElasticsearchLogClient factory to be passed as option for DurableElasticsearchSink
The `ElasticsearchLogClient` uses internally `DynamicResponse` object that is not usable in when application is compiled to native code. This is usually the case with UWP apps as those are compiled to native code by default.

So this changes allows `ElasticsearchLogClient` to be overridded with custom implementation when `ElasticsearchLogClient` cannot be used.

For more information about the issue with `DynamicResponse` and native code compilation see https://github.com/elastic/elasticsearch-net/issues/4389